Reference

PoincareInvariants — Module

PoincareInvariants

A Julia library for the computation of Poincaré integral invariants.

PoincareInvariants.AbstractPoincareInvariant — Type

AbstractPoincareInvariant

supertype of FirstPoincareInvariant and SecondPoincareInvariant.

PoincareInvariants.compute! — Function

compute!(pinv::AbstractPoincareInvariant, args...)

computes a Poincaré invariant.

PoincareInvariants.getpoints — Function

getpoints(pinv::AbstractPoincareInvariant)

returns points on which to evaluate the phase space line or surface parameterisation so as to compute! pinv.

PoincareInvariants.getpointnum — Function

getpointnum(pinv::AbstractPoincareInvariant)

returns number of points to sample in phase space to compute! pinv.

PoincareInvariants.getdim — Function

getdim(pinv::AbstractPoincareInvariant)

returns dimension of phase space to compute! pinv in.

PoincareInvariants.getform — Function

getform(pinv::AbstractPoincareInvariant)

get invariant one- or two-form.